refactored structure of client to make it have a components folder
